refactor: update sgxInstance assignment in NewPublicInput and rename test package
