+++ tests/aggregate_test.go
@@ -1,0 +1,7 @@
+package tests
+
+import "testing"
+
+func TestAggregate(t *testing.T) {
+
+}
