+++ tests/batch_test.go
@@ -1,4 +1,4 @@
-package test
+package tests
 
 import (
 	"bytes"
